+++ .eslintrc.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended'],
+  // parser: '@typescript-eslint/parser',
+  // plugins: ['@typescript-eslint'],
+  root: true,
+};
