+++ LICENSE.md
@@ -1,0 +1,9 @@
+MapTiler Weather JS Module
+
+Copyright © 2023 MapTiler AG. All rights reserved.
+
+The software and files (collectively “Software”) in this repository are licensed for use only with MapTiler service(s). 
+
+For the license terms, please reference  [MapTiler General Terms and Conditions](https://www.maptiler.com/terms/) which incorporate MapTiler Weather JS Module Product [Terms (collectively “Terms”) and Privacy Policy at Privacy policy](https://www.maptiler.com/privacy-policy). 
+
+This license allows users with an active MapTiler account to modify and integrate authorized portions of the Software for use with the relevant MapTiler service(s) in accordance with the MapTiler Terms. This license terminates automatically if a user no longer maintains a MapTiler account or their usage breaches MapTiler Terms.
