+++ demos/helpers.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-    <title>Geolocate</title>
+    <title>WebGL</title>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <style>
@@ -12,6 +12,8 @@
       }
     </style>
 
+
+
     <!-- Leaflet -->
     <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.3/dist/leaflet.css" />
     <script src="https://unpkg.com/leaflet@1.9.3/dist/leaflet.js"></script>
@@ -20,23 +22,38 @@
     <script src="https://cdn.maptiler.com/maptiler-sdk-js/latest/maptiler-sdk.umd.js"></script>
     <link href="https://cdn.maptiler.com/maptiler-sdk-js/latest/maptiler-sdk.css" rel="stylesheet" />
 
+    <!-- Leaflet plugin for MapTiler SDK Layers -->
+    <script src="../build/leaflet-maptilersdk.umd.js"></script>
 </head>
 
 <body>
+<!-- The map container -->
 <div id="map"></div>
 
-<script src="../leaflet-maptilersdk.js"></script>
+
 <script>
 
-// Leaflet still requires to set the view, 
-// but this is going to be superseded by the `geolocate: true` options
-const map = L.map('map').setView([0, 0], 0);
 
-const mtLayer = L.maptilerLayer({
-  style: L.maptilerStyle.BASIC,
-  apiKey: "YOUR_MAPTILER_API_KEY",
-  geolocate: true,
-}).addTo(map);
+
+// Create a MapTiler Layer inside Leaflet
+// const mtLayer = L.maptilerLayer({
+//   apiKey: "PhsGJ2SdW2MhmzzsT8zU",
+//   style: L.MaptilerStyle.DATAVIZ.DARK,
+// }).addTo(map);
+
+// mtLayer.addPolyline({data: "74003ba7-215a-4b7e-8e26-5bbe3aa70b05"})
+
+(async () => {
+  const map = L.map('map').setView([46.3796, 6.1518], 13);
+  const mtLayer = await L.maptilerLayerAsync({
+    apiKey: "PhsGJ2SdW2MhmzzsT8zU",
+    style: L.MaptilerStyle.DATAVIZ.DARK,
+  });
+
+  mtLayer.addTo(map);
+})()
+
+
 
 </script>
 </body>
