--- examples/basic.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>WebGL</title>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <style>
-      html, body, #map {
-        width: 100%;
-        height: 100%;
-        margin: 0;
-      }
-    </style>
-
-    <!-- Leaflet -->
-    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.3/dist/leaflet.css" />
-    <script src="https://unpkg.com/leaflet@1.9.3/dist/leaflet.js"></script>
-
-    <!-- MapTiler SDK -->
-    <script src="https://cdn.maptiler.com/maptiler-sdk-js/latest/maptiler-sdk.umd.js"></script>
-    <link href="https://cdn.maptiler.com/maptiler-sdk-js/latest/maptiler-sdk.css" rel="stylesheet" />
-
-</head>
-
-<body>
-<div id="map"></div>
-
-<script src="../leaflet-maptilersdk.js"></script>
-<script>
-const map = L.map('map').setView([40.7468, -73.98775], 13);
-
-L.marker([40.7468, -73.98775])
-  .bindPopup("Hello <b>Leaflet with MapTiler SDK</b>!<br>Whoa, it works!")
-  .addTo(map)
-  .openPopup();
-
-const mtLayer = L.maptilerLayer({
-  // get your own MapTiler token at https://cloud.maptiler.com/ or use MapBox style
-  apiKey: "YOUR_MAPTILER_API_KEY",
-}).addTo(map);
-
-
-
-
-
-</script>
-</body>
-</html>
