--- examples/cluster.html
@@ -1,60 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Cluster</title>
-    <meta charset="utf-8">
-    <style>
-    
-    html, body, #map {
-        width: 100%;
-        height: 100%;
-        margin: 0;
-    }
-    
-    </style>
-
-    <!-- Leaflet -->
-    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.3/dist/leaflet.css" />
-    <script src="https://unpkg.com/leaflet@1.9.3/dist/leaflet.js"></script>
-
-    <!-- MapTiler SDK -->
-    <script src="https://cdn.maptiler.com/maptiler-sdk-js/latest/maptiler-sdk.umd.js"></script>
-    <link href="https://cdn.maptiler.com/maptiler-sdk-js/latest/maptiler-sdk.css" rel="stylesheet" />
-
-    <!-- Leaflet.MarkerCluster -->
-    <script src='https://cdnjs.cloudflare.com/ajax/libs/leaflet.markercluster/1.5.3/leaflet.markercluster.js'></script>
-    <link href='https://cdnjs.cloudflare.com/ajax/libs/leaflet.markercluster/1.5.3/MarkerCluster.css' rel='stylesheet' />
-    <link href='https://cdnjs.cloudflare.com/ajax/libs/leaflet.markercluster/1.5.3/MarkerCluster.Default.css' rel='stylesheet' />
-
-    <!-- test data -->
-    <script src="https://www.mapbox.com/mapbox.js/assets/data/realworld.388.js"></script>
-</head>
-
-<body>
-<div id="map"></div>
-
-<script src="../leaflet-maptilersdk.js"></script>
-<script>
-
-const map = L.map('map', {maxZoom: 22}).setView([-37.821, 175.219], 16);
-
-const mtLayer = L.maptilerLayer({
-  // get your own MapTiler token at https://cloud.maptiler.com/ or use MapBox style
-  apiKey: "YOUR_MAPTILER_API_KEY",
-}).addTo(map);
-
-const markers = L.markerClusterGroup();
-
-for (let i = 0; i < addressPoints.length; i++) {
-    const a = addressPoints[i];
-    const title = a[2];
-    const marker = L.marker(new L.LatLng(a[0], a[1]), { title: title });
-    marker.bindPopup(title);
-    markers.addLayer(marker);
-}
-
-map.addLayer(markers);
-
-</script>
-</body>
-</html>
