--- examples/events.html
@@ -1,87 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Events</title>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <style>
-      html, body, #map {
-        width: 100%;
-        height: 100%;
-        margin: 0;
-      }
-    </style>
-
-    <!-- Leaflet -->
-    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.3/dist/leaflet.css" />
-    <script src="https://unpkg.com/leaflet@1.9.3/dist/leaflet.js"></script>
-
-    <!-- MapTiler SDK -->
-    <script src="https://cdn.maptiler.com/maptiler-sdk-js/latest/maptiler-sdk.umd.js"></script>
-    <link href="https://cdn.maptiler.com/maptiler-sdk-js/latest/maptiler-sdk.css" rel="stylesheet" />
-  </head>
-  <body>
-  <div id="map"></div>
-  <script src="../leaflet-maptilersdk.js"></script>
-  <script>
-  var leafletMap = L.map('map').setView([38.912753, -77.032194], 4);
-
-  L.marker([38.912753, -77.032194])
-    .bindPopup("Hello <b>Leaflet GL</b>!<br>Whoa, it works!")
-    .addTo(leafletMap)
-    .openPopup();
-
-  const mtLayer = L.maptilerLayer({
-    padding: 0.1,
-    // get your own MapTiler token at https://cloud.maptiler.com/ or use MapBox style
-    style: "streets-v2",
-    apiKey: "YOUR_MAPTILER_API_KEY",
-    interactive: true
-  }).addTo(leafletMap);
-
-  const maptilerMap = mtLayer.getMaptilerSDKMap();
-
-  maptilerMap.on('load', () => {
-    console.log('MapTiler SDK map loaded');
-
-    // let's see events on MapTiler SDK map
-    maptilerMap.on('mousemove', () => { console.log('MapTiler SDK mousemove') });
-    maptilerMap.on('mouseenter', () => { console.log('MapTiler SDK mouseenter') });
-    maptilerMap.on('mouseout', () => { console.log('MapTiler SDK mouseout') });
-    maptilerMap.on('mouseleave', () => { console.log('MapTiler SDK mouseleave') });
-    maptilerMap.on('mouseover', () => { console.log('MapTiler SDK mouseover') });
-
-    // let's add some layer and fire events on it
-
-    maptilerMap.addSource('states', {
-      'type': 'geojson',
-      'data': 'https://docs.mapbox.com/mapbox-gl-js/assets/us_states.geojson'
-    });
-    maptilerMap.addLayer({
-      'id': 'state-fills',
-      'type': 'fill',
-      'source': 'states',
-      'layout': {},
-      'paint': {
-        'fill-color': '#627BC1',
-        'fill-opacity': ['case', ['boolean', ['feature-state', 'hover'], false], 1, 0.5 ]
-      }
-    });
-
-    maptilerMap.on('mouseenter', 'state-fills', (e) => {console.log('state-fills mouseenter', e) });
-    maptilerMap.on('mousemove', 'state-fills', (e) => {console.log('state-fills mousemove', e) });
-    maptilerMap.on('mouseout', 'state-fills', (e) => {console.log('state-fills mouseout', e) });
-    maptilerMap.on('mouseleave', 'state-fills', (e) => {console.log('state-fills mouseleave', e) });
-
-  });
-
-  // now let's see on leaflet map events
-  // SPOILER: they are works
-  leafletMap.on('mousemove', () => { console.log('LEAFLET mousemove') });
-  leafletMap.on('mouseenter', () => { console.log('LEAFLET mouseenter') });
-  leafletMap.on('mouseout', () => { console.log('LEAFLET mouseout') });
-  leafletMap.on('mouseleave', () => { console.log('LEAFLET mouseleave') });
-  leafletMap.on('mouseover', () => { console.log('LEAFLET mouseover') });
-    </script>
-  </body>
-</html>
