--- examples/language.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Language</title>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <style>
-      html, body, #map {
-        width: 100%;
-        height: 100%;
-        margin: 0;
-      }
-    </style>
-
-    <!-- Leaflet -->
-    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.3/dist/leaflet.css" />
-    <script src="https://unpkg.com/leaflet@1.9.3/dist/leaflet.js"></script>
-
-    <!-- MapTiler SDK -->
-    <script src="https://cdn.maptiler.com/maptiler-sdk-js/latest/maptiler-sdk.umd.js"></script>
-    <link href="https://cdn.maptiler.com/maptiler-sdk-js/latest/maptiler-sdk.css" rel="stylesheet" />
-
-</head>
-
-<body>
-<div id="map"></div>
-
-<script src="../leaflet-maptilersdk.js"></script>
-<script>
-
-// Leaflet still requires to set the view, 
-const map = L.map('map').setView([42, 40], 4);
-
-const mtLayer = L.maptilerLayer({
-  style: L.maptilerStyle.STREETS.DARK,
-  apiKey: "YOUR_MAPTILER_API_KEY",
-  language: L.maptilerLanguage.JAPANESE,
-}).addTo(map);
-
-</script>
-</body>
-</html>
