--- examples/overlay.html
@@ -1,53 +1,0 @@
-<html>
-<head>
-	<title>Overlay</title>
-
-  <!-- Leaflet -->
-  <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.3/dist/leaflet.css" />
-  <script src="https://unpkg.com/leaflet@1.9.3/dist/leaflet.js"></script>
-
-  <!-- MapTiler SDK -->
-  <script src="https://cdn.maptiler.com/maptiler-sdk-js/latest/maptiler-sdk.umd.js"></script>
-  <link href="https://cdn.maptiler.com/maptiler-sdk-js/latest/maptiler-sdk.css" rel="stylesheet" />
-  
-  <script src="../leaflet-maptilersdk.js"></script>
-  <style>
-    html, body, #map {
-      width: 100%;
-      height: 100%;
-      margin: 0;
-    }
-  </style>
-</head>
-<body>
-
-	<div id="map"></div>
-
-	<script>
-    const map = new L.Map('map');
-
-		const mtLayer = L.maptilerLayer({
-      // get your own MapTiler token at https://cloud.maptiler.com/ or use MapBox style
-      style: L.maptilerStyle.DATAVIZ.LIGHT,
-      apiKey: "YOUR_MAPTILER_API_KEY",
-    }).addTo(map);
-
-		const	bounds = new L.LatLngBounds(
-			new L.LatLng(40.71222,-74.22655),
-			new L.LatLng(40.77394,-74.12544));
-
-		map.fitBounds(bounds);
-
-		const overlay = new L.ImageOverlay("https://www.lib.utexas.edu/maps/historical/newark_nj_1922.jpg", bounds, {
-			opacity: 0.5,
-			interactive: true,
-			attribution: '&copy; A.B.B Corp.'
-		});
-
-		map.addLayer(overlay);
-		overlay.on('dblclick',function (e) {
-			console.log('Double click on image.');
-		});
-	</script>
-</body>
-</html>
