+++ vite.config-es.js
@@ -1,0 +1,36 @@
+import { resolve } from 'path';
+import { defineConfig } from 'vite';
+
+const isProduction = process.env.NODE_ENV === "production";
+
+export default defineConfig({
+  mode: isProduction ? "production" : "development",
+  build: {
+    minify: isProduction,
+
+    sourcemap: !isProduction,
+    lib: {
+      // Could also be a dictionary or array of multiple entry points
+      entry: resolve(__dirname, 'src/leaflet-maptilersdk.js'),
+      name: 'leafletmaptilersdk',
+      // the proper extensions will be added
+      fileName: (format, entryName) => `${entryName}.js`,
+      formats: ['es'],
+    },
+    
+    rollupOptions: {
+      // make sure to externalize deps that shouldn't be bundled
+      // into your library
+      external: [
+        "leaflet",
+        "@maptiler/sdk",
+      ],
+      output: {
+        // Provide global variables to use in the UMD build
+        // for externalized deps
+        globals: {},
+      },
+    },
+  },
+  plugins: [],
+})
